refactor(ctr): store item embeddings in bf16 (#1237)

+++ dataset/dataset.go
@@ -69,7 +69,7 @@ type CTRSplit interface {
 	CountNegative() int
 	GetIndex() UnifiedIndex
 	GetTarget(i int) float32
-	Get(i int) ([]int32, []float32, [][]float32, float32)
+	Get(i int) ([]int32, []float32, [][]uint16, float32)
 	GetItemEmbeddingDim() []int
 	GetItemEmbeddingIndex() *Index
 }
